requirements.txt
streamlit==1.23.0
scikit-learn==1.2.2
joblib==1.2.0
matplotlib==3.7.1
numpy==1.22.4
pandas==1.5.3
pip-tools==6.13.0
scipy==1.10.1
seaborn==0.12.2
sklearn-pandas==2.2.0
tqdm==4.65.0
